deyisiklik

+++ src/products/products.service.ts
@@ -16,27 +16,40 @@ export class ProductsService {
   }
 
   // Belirli bir ürünü almak için
-  findOne(id: string): Promise<Product> {
-    return this.productsRepository.findOne({ where : {id} });
+  async findOne(id: string): Promise<Product> {
+    const product = await this.productsRepository.findOne({ where: { id } });
+    if (!product) {
+      throw new Error('Product not found');
+    }
+    return product;
   }
 
   // Yeni bir ürün eklemek için
- async create(product: Product): Promise<Product> {
-    console.log("product",product)
-    const newProduct= await new Product(product)
-    console.log("newProduct",newProduct)
-
-    return this.productsRepository.save(newProduct);
+  async create(product: Product): Promise<Product> {
+    console.log("product", product);
+    return this.productsRepository.save(product);
   }
 
   // Belirli bir ürünü güncellemek için
-  async update(id: string, product: Product): Promise<Product> {
-    await this.productsRepository.update(id, product);
-    return this.findOne(id);
+  async update(id: string, product: Partial<Product>): Promise<Product> {
+    const existingProduct = await this.findOne(id);
+
+    if (!existingProduct) {
+      throw new Error('Product not found');
+    }
+
+    // Verilen product data ile güncellemeyi yap
+    Object.assign(existingProduct, product);
+
+    return this.productsRepository.save(existingProduct);
   }
 
   // Belirli bir ürünü silmek için
   async remove(id: string): Promise<void> {
+    const product = await this.findOne(id);
+    if (!product) {
+      throw new Error('Product not found');
+    }
     await this.productsRepository.delete(id);
   }
 }
